requirements.txt
django
psycopg2-binary
djangorestframework
pika
django-filter